+++ src/modules/mentee/dto/create-mentee.dto.ts
@@ -1,0 +1,9 @@
+/* eslint-disable prettier/prettier */
+import { IsNotEmpty, IsString } from 'class-validator';
+import { BaseDto } from 'src/base/dto/base.dto';
+
+export class CreateMenteeDto extends BaseDto {
+  @IsString()
+  @IsNotEmpty()
+  mentee_name: string;
+}
