--- src/modules/mentee/dto/create.mentee.dto.ts
@@ -1,8 +1,0 @@
-import { IsEmail, IsNotEmpty, IsString } from "class-validator"
-import { BaseDto } from "src/base/dto/base.dto"
-
-export class CreateMenteeDto extends BaseDto{
-    @IsString()
-    @IsNotEmpty()
-    mentee_name: string
-}
