+++ src/modules/mentor/dto/edit-mentor.dto.ts
@@ -1,0 +1,3 @@
+import { CreateMentorDto } from './create-mentor.dto';
+
+export class EditMentorDto extends CreateMentorDto {}
